// src/app/models/customer.model.ts
export class CustomerData {
    Amsidnr: string;
    Street: string;
    ModifyDate: string;
    Lastname: string;
    AgentNumber: string;
    Bank1: string;
    Bank2: string;
    Bic1: string;
    Bic2: string;
    Email: string;
    Birthdate: string;
    Gender: string;
    Iban1: string;
    Iban2: string;
    DisplayName: string;
    City: string;
    ZipCode: string;
    Telefon1: string;
    Telefon2: string;
    Email1: string;
    Email2: string;
    ShowName: string;
    Firstname: string
}